+++ src/pages/index/ccc.js
@@ -1,0 +1,916 @@
+import React, { useState } from 'react'
+import './pages/index/App.css'
+
+function App() {
+  return (
+    <>
+      <div>
+        <div className="container">
+          <div className="nav ">
+            <div className="icon-s-none ">
+              <a href="#">
+                {/*?xml version="1.0" encoding="iso-8859-1"?*/}
+                {/* Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In . SVG Version: 6.00 Build 0)  */}
+                <svg
+                  className="nav-icon"
+                  version="1.1"
+                  id="Layer_1"
+                  xmlns="http://www.w3.org/2000/svg"
+                  xmlnsXlink="http://www.w3.org/1999/xlink"
+                  x="0px"
+                  y="0px"
+                  viewBox="0 0 469.333 469.333"
+                  style={{ enableBackground: 'new 0 0 469.333 469.333' }}
+                  xmlSpace="preserve"
+                >
+                  <g>
+                    <g>
+                      <g>
+                        <path
+                          d="M53.333,106.667H416c29.417,0,53.333-23.927,53.333-53.333S445.417,0,416,0H53.333C23.917,0,0,23.927,0,53.333
+				S23.917,106.667,53.333,106.667z M53.333,21.333H416c17.646,0,32,14.354,32,32c0,17.646-14.354,32-32,32H53.333
+				c-17.646,0-32-14.354-32-32C21.333,35.687,35.687,21.333,53.333,21.333z"
+                        />
+                        <path
+                          d="M416,181.333H53.333C23.917,181.333,0,205.26,0,234.667S23.917,288,53.333,288H416c29.417,0,53.333-23.927,53.333-53.333
+				S445.417,181.333,416,181.333z M416,266.667H53.333c-17.646,0-32-14.354-32-32s14.354-32,32-32H416c17.646,0,32,14.354,32,32
+				S433.646,266.667,416,266.667z"
+                        />
+                        <path
+                          d="M416,362.667H53.333C23.917,362.667,0,386.594,0,416s23.917,53.333,53.333,53.333H416
+				c29.417,0,53.333-23.927,53.333-53.333S445.417,362.667,416,362.667z M416,448H53.333c-17.646,0-32-14.354-32-32
+				c0-17.646,14.354-32,32-32H416c17.646,0,32,14.354,32,32C448,433.646,433.646,448,416,448z"
+                        />
+                      </g>
+                    </g>
+                  </g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                </svg>
+              </a>
+            </div>
+            <div className="logo">
+              <a href="#/">
+                <img src="./svg/LOGO.svg" alt="" />
+              </a>
+            </div>
+            <div className="navbar">
+              <li>
+                <div className=" nav-three-gp">
+                  <a href="#/">
+                    <div className="three-gp-flex">
+                      <div>
+                        <svg
+                          className="nav-icon"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 101.01 101"
+                        >
+                          <defs>
+                            <style
+                              dangerouslySetInnerHTML={{
+                                __html:
+                                  '\n\n                                            ',
+                              }}
+                            />
+                          </defs>
+                          <g id="圖層_2" data-name="圖層 2">
+                            <g id="圖層_1-2" data-name="圖層 1">
+                              <path
+                                className="cls-1"
+                                d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                              />
+                              <path
+                                className="cls-1"
+                                d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                              />
+                            </g>
+                          </g>
+                        </svg>
+                      </div>
+                      <div>
+                        <div>店家地圖</div>
+                        <div>store map</div>
+                      </div>
+                    </div>
+                  </a>
+                </div>
+              </li>
+              <li>
+                <div className=" nav-three-gp">
+                  <a href="#/">
+                    <div className="three-gp-flex">
+                      <div>
+                        <svg
+                          className="nav-icon"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 101.01 101"
+                        >
+                          <defs>
+                            <style
+                              dangerouslySetInnerHTML={{
+                                __html:
+                                  '\n\n                                            ',
+                              }}
+                            />
+                          </defs>
+                          <g id="圖層_2" data-name="圖層 2">
+                            <g id="圖層_1-2" data-name="圖層 1">
+                              <path
+                                className="cls-1"
+                                d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                              />
+                              <path
+                                className="cls-1"
+                                d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                              />
+                            </g>
+                          </g>
+                        </svg>
+                      </div>
+                      <div>
+                        <div>咖啡介紹</div>
+                        <div>store map</div>
+                      </div>
+                    </div>
+                  </a>
+                </div>
+              </li>
+              <li>
+                <div className=" nav-three-gp">
+                  <a href="#/">
+                    <div className="three-gp-flex">
+                      <div>
+                        <svg
+                          className="nav-icon"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 101.01 101"
+                        >
+                          <defs>
+                            <style
+                              dangerouslySetInnerHTML={{
+                                __html:
+                                  '\n\n                                            ',
+                              }}
+                            />
+                          </defs>
+                          <g id="圖層_2" data-name="圖層 2">
+                            <g id="圖層_1-2" data-name="圖層 1">
+                              <path
+                                className="cls-1"
+                                d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                              />
+                              <path
+                                className="cls-1"
+                                d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                              />
+                            </g>
+                          </g>
+                        </svg>
+                      </div>
+                      <div>
+                        <div>全部商品</div>
+                        <div>store map</div>
+                      </div>
+                    </div>
+                  </a>
+                </div>
+              </li>
+              <li>
+                <div className=" nav-three-gp  d-play-none">
+                  <a href="#/">
+                    <div className="three-gp-flex ">
+                      <div>
+                        <svg
+                          className="nav-icon"
+                          xmlns="http://www.w3.org/2000/svg"
+                          viewBox="0 0 101.01 101"
+                        >
+                          <defs>
+                            <style
+                              dangerouslySetInnerHTML={{
+                                __html:
+                                  '\n\n                                            ',
+                              }}
+                            />
+                          </defs>
+                          <g id="圖層_2" data-name="圖層 2">
+                            <g id="圖層_1-2" data-name="圖層 1">
+                              <path
+                                className="cls-1"
+                                d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                              />
+                              <path
+                                className="cls-1"
+                                d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                              />
+                            </g>
+                          </g>
+                        </svg>
+                      </div>
+                      <div>
+                        <div>關於我們</div>
+                        <div>store map</div>
+                      </div>
+                    </div>
+                  </a>
+                </div>
+              </li>
+            </div>
+            <div className="loginbar">
+              <div className="three-gp-flex">
+                <a href="#/">
+                  <svg
+                    className="nav-icon-2"
+                    xmlns="http://www.w3.org/2000/svg"
+                    viewBox="0 0 101.01 101"
+                  >
+                    <defs>
+                      <style
+                        dangerouslySetInnerHTML={{
+                          __html: '\n\n                                ',
+                        }}
+                      />
+                    </defs>
+                    <g id="圖層_2" data-name="圖層 2">
+                      <g id="圖層_1-2" data-name="圖層 1">
+                        <path
+                          className="cls-1"
+                          d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                        />
+                        <path
+                          className="cls-1"
+                          d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                        />
+                      </g>
+                    </g>
+                  </svg>
+                  登入/註冊
+                </a>
+              </div>
+              <div className="three-gp-flex">
+                <a href="#/">
+                  <svg
+                    className="nav-icon-2"
+                    xmlns="http://www.w3.org/2000/svg"
+                    viewBox="0 0 101.01 101"
+                  >
+                    <defs>
+                      <style
+                        dangerouslySetInnerHTML={{
+                          __html: '\n\n                                ',
+                        }}
+                      />
+                    </defs>
+                    <g id="圖層_2" data-name="圖層 2">
+                      <g id="圖層_1-2" data-name="圖層 1">
+                        <path
+                          className="cls-1"
+                          d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                        />
+                        <path
+                          className="cls-1"
+                          d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                        />
+                      </g>
+                    </g>
+                  </svg>
+                  店家專區
+                </a>
+              </div>
+              <div>
+                <a href="#">
+                  <svg
+                    className="nav-icon"
+                    xmlns="http://www.w3.org/2000/svg"
+                    viewBox="0 0 101.01 101"
+                  >
+                    <defs>
+                      <style
+                        dangerouslySetInnerHTML={{
+                          __html: '\n\n                                ',
+                        }}
+                      />
+                    </defs>
+                    <g id="圖層_2" data-name="圖層 2">
+                      <g id="圖層_1-2" data-name="圖層 1">
+                        <path
+                          className="cls-1"
+                          d="M63,27.66A12.53,12.53,0,1,0,50.5,40.18,12.52,12.52,0,0,0,63,27.66ZM41,27.66a9.56,9.56,0,1,1,9.55,9.56A9.56,9.56,0,0,1,41,27.66Z"
+                        />
+                        <path
+                          className="cls-1"
+                          d="M100.91,99,88.79,67.21a1.49,1.49,0,0,0-1.39-1H64.46l.42-.68c8.83-14.61,13.27-27.39,13.2-38a27.58,27.58,0,0,0-55.16,0c-.06,10.61,4.38,23.39,13.21,38l.42.68H13.61a1.48,1.48,0,0,0-1.39,1L.1,99A1.49,1.49,0,0,0,1,100.9a1.44,1.44,0,0,0,.54.1h98A1.49,1.49,0,0,0,101,99.52a1.62,1.62,0,0,0-.1-.54ZM26,27.66a24.55,24.55,0,1,1,49.09,0h0C75.18,48.73,55.48,74.72,50.5,81c-5-6.23-24.67-32.21-24.54-53.29ZM49.37,84.25a1.48,1.48,0,0,0,2.09.17,1.37,1.37,0,0,0,.17-.17A138.71,138.71,0,0,0,62.57,69H86.29l1.89,5.17L41.83,90.67,25.31,69.22H38.44a157.35,157.35,0,0,0,10.93,15ZM41.78,93.8h0l27.65-9.65L80.17,97.82H30.25ZM14.7,69.22h6.87L38.94,91.79,21.4,98H3.71ZM83.91,98l-11.53-15,16.86-6,8,21Z"
+                        />
+                      </g>
+                    </g>
+                  </svg>
+                </a>
+              </div>
+            </div>
+            <div className="icon-s-none ">
+              <a href="#/">
+                {/*?xml version="1.0" encoding="iso-8859-1"?*/}
+                {/* Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In . SVG Version: 6.00 Build 0)  */}
+                <svg
+                  className="nav-icon"
+                  version="1.1"
+                  id="Capa_1"
+                  xmlns="http://www.w3.org/2000/svg"
+                  xmlnsXlink="http://www.w3.org/1999/xlink"
+                  x="0px"
+                  y="0px"
+                  viewBox="0 0 450.391 450.391"
+                  style={{ enableBackground: 'new 0 0 450.391 450.391' }}
+                  xmlSpace="preserve"
+                >
+                  <g>
+                    <g>
+                      <g>
+                        <path
+                          d="M143.673,350.322c-25.969,0-47.02,21.052-47.02,47.02c0,25.969,21.052,47.02,47.02,47.02
+                                    c25.969,0,47.02-21.052,47.02-47.02C190.694,371.374,169.642,350.322,143.673,350.322z M143.673,423.465
+                                    c-14.427,0-26.122-11.695-26.122-26.122c0-14.427,11.695-26.122,26.122-26.122c14.427,0,26.122,11.695,26.122,26.122
+                                    C169.796,411.77,158.1,423.465,143.673,423.465z"
+                        />
+                        <path
+                          d="M342.204,350.322c-25.969,0-47.02,21.052-47.02,47.02c0,25.969,21.052,47.02,47.02,47.02s47.02-21.052,47.02-47.02
+                                    C389.224,371.374,368.173,350.322,342.204,350.322z M342.204,423.465c-14.427,0-26.122-11.695-26.122-26.122
+                                    c0-14.427,11.695-26.122,26.122-26.122s26.122,11.695,26.122,26.122C368.327,411.77,356.631,423.465,342.204,423.465z"
+                        />
+                        <path
+                          d="M448.261,76.037c-2.176-2.377-5.153-3.865-8.359-4.18L99.788,67.155L90.384,38.42
+                                    C83.759,19.211,65.771,6.243,45.453,6.028H10.449C4.678,6.028,0,10.706,0,16.477s4.678,10.449,10.449,10.449h35.004
+                                    c11.361,0.251,21.365,7.546,25.078,18.286l66.351,200.098l-5.224,12.016c-5.827,15.026-4.077,31.938,4.702,45.453
+                                    c8.695,13.274,23.323,21.466,39.184,21.943h203.233c5.771,0,10.449-4.678,10.449-10.449c0-5.771-4.678-10.449-10.449-10.449
+                                    H175.543c-8.957-0.224-17.202-4.936-21.943-12.539c-4.688-7.51-5.651-16.762-2.612-25.078l4.18-9.404l219.951-22.988
+                                    c24.16-2.661,44.034-20.233,49.633-43.886l25.078-105.012C450.96,81.893,450.36,78.492,448.261,76.037z M404.376,185.228
+                                    c-3.392,15.226-16.319,26.457-31.869,27.69l-217.339,22.465L106.58,88.053l320.261,4.702L404.376,185.228z"
+                        />
+                      </g>
+                    </g>
+                  </g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                </svg>
+              </a>
+            </div>
+          </div>
+          <div className="herro-section">
+            <div className="onepage-pic">
+              <img src="./img/herrosection.jpg" alt="no-pic" />
+            </div>
+            <div className="img-s-none">
+              <img src="./img/phone-herrosection.jpg" alt="" />
+            </div>
+            <div className="herro-text">
+              <div>
+                {/*?xml version="1.0" encoding="iso-8859-1"?*/}
+                {/* Generator: Adobe Illustrator 19.0.0, SVG Export Plug-In . SVG Version: 6.00 Build 0)  */}
+                <svg
+                  className="nav-icon"
+                  version="1.1"
+                  id="Capa_1"
+                  xmlns="http://www.w3.org/2000/svg"
+                  xmlnsXlink="http://www.w3.org/1999/xlink"
+                  x="0px"
+                  y="0px"
+                  viewBox="0 0 512 512"
+                  style={{ enableBackground: 'new 0 0 512 512' }}
+                  xmlSpace="preserve"
+                >
+                  <g>
+                    <g>
+                      <polygon
+                        points="447.894,263.57 275.989,435.474 275.989,0 236.011,0 236.011,435.474 64.106,263.57 35.838,291.838 256,512 
+			476.162,291.838 		"
+                      />
+                    </g>
+                  </g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                  <g></g>
+                </svg>
+              </div>
+              <div className="d-flex-columm">
+                <div>旅行‧寧靜‧生活</div>
+                <div>Mercurius Coffee</div>
+              </div>
+            </div>
+            <div className="onepage-rectangle" />
+            <div className="herro-button-left">
+              <a href="#/">尋找咖啡</a>{' '}
+            </div>
+            <div className="herro-button-right">
+              <a href="#/">尋找店家</a>{' '}
+            </div>
+          </div>
+          <div className="two-section">
+            <div className="two-colorship">
+              <img src="./img/colorshlip.png" />{' '}
+            </div>
+            <div className=" twopage-pic">
+              <img src="./img/twosection.jpg" alt="no-pic" />
+            </div>
+            <div className=" phone-twopage-pic">
+              <img src="./img/phone-twosection.jpg" alt="no-pic" />
+            </div>
+            <div className="cardtwo">
+              <div className="fontsize-34">我們相遇在世界的角落</div>
+              <div className="mt-30">
+                <div className="fontsize-34 px-15">在台灣再度重逢</div>
+                <div className="fontsize-34 px-15">在你手中</div>
+                <div className="fontsize-34 px-15">為你研磨最好的一品</div>
+              </div>
+              <div className="button-blue-left">
+                <a href="#/">尋找咖啡</a>
+              </div>
+              <div className="button-blue-right">
+                <a href="#/">尋找店家</a>
+              </div>
+            </div>
+            <div className="twopage-rectangle" />
+          </div>
+          <div className="phone-long-colorship">
+            <img src="./img/long-colorship.png" alt="no-pic" />
+          </div>
+          <div className="three-section">
+            <div className="threepage-pic">
+              <img src="./img/threesection.jpg" alt="no-pic" />
+            </div>
+            <div className="phone-three-pic">
+              <img src="./img/phone-threesection.jpg" alt="" />
+            </div>
+            <div className="three-content">
+              <div className="fontsize60">
+                <div className="py-15">在車水馬龍的都會中，</div>
+                <div className="py-15">你渴求著片刻的喘息...</div>
+              </div>
+              <div className="fontsize16 py-15">
+                我們將守著那寧靜的角落，將一抹芬芳贈與給你
+              </div>
+            </div>
+            <div className=" fontsize-34 newproduct ">
+              New product listing
+              <div className="border-bottom-white"></div>
+            </div>
+            <div className="button-black fontsize-20">
+              <a href="#/"> ALL</a>
+              <div>
+                <img src alt="" />{' '}
+              </div>
+            </div>
+            <div className="box">
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超d999999999ddddcsssdddcdsss好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box1">
+                <div className="box1-pic">
+                  <img src="./img/productcard.jpg" alt="no-pic" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="button03">
+            <div>
+              <button>左</button>
+            </div>
+            <div>
+              <button>右</button>
+            </div>
+          </div>
+          <div className="long-colorship">
+            <img src="./img/long-colorship.png" alt="no-pic" />
+          </div>
+          <div className="container-1">
+            <div className="best-selling-group">
+              <div className="fontsize-34 ">
+                Best-selling goods
+                <div className="border-bottom-blue"> </div>
+              </div>
+              <div className=" button-blue-bg fontsize-20">
+                <a href="#/"> ALL </a>
+                <div>
+                  <img src alt="" />{' '}
+                </div>
+              </div>
+            </div>
+            <div className="best-selling-item">
+              <div className="box2">
+                <div className="box2-group">
+                  <div className="box2-pic">
+                    <img src="./img/productcard.jpg" alt="" />
+                  </div>
+                  <div className="item-1">
+                    <div className="item-2">
+                      <div className="fontsize-20">精選咖啡豆</div>
+                      <div>svg</div>
+                    </div>
+                    <div>
+                      超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                    </div>
+                    <div className="card-button">
+                      <a href="#/"> 購買</a>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div className="box3">
+                <div className="box2-pic">
+                  <img src="./img/productcard.jpg" alt="" />
+                </div>
+                <div className="item-1">
+                  <div className="item-2">
+                    <div className="fontsize-20">精選咖啡豆</div>
+                    <div>svg</div>
+                  </div>
+                  <div>
+                    超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                  </div>
+                  <div className="card-button">
+                    <a href="#/"> 購買</a>
+                  </div>
+                </div>
+              </div>
+              <div className="box4">
+                <div className="aaa">
+                  <div className="box2-pic">
+                    <img src="./img/productcard.jpg" alt="" />
+                  </div>
+                  <div className="item-1">
+                    <div className="item-2">
+                      <div className="fontsize-20">精選咖啡豆</div>
+                      <div>svg</div>
+                    </div>
+                    <div>
+                      超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                    </div>
+                    <div className="card-button">
+                      <a href="#/"> 購買</a>
+                    </div>
+                  </div>
+                </div>
+              </div>
+              <div className="box5">
+                <div className="aaa">
+                  <div className="box2-pic">
+                    <img src="./img/productcard.jpg" alt="" />
+                  </div>
+                  <div className="item-1">
+                    <div className="item-2">
+                      <div className="fontsize-20">精選咖啡豆</div>
+                      <div>svg</div>
+                    </div>
+                    <div>
+                      超好喝的超好喝的超好喝的超好喝的超好喝的超好喝的=超好喝的超好喝的超好喝的
+                    </div>
+                    <div className="card-button">
+                      <a href="#/"> 購買</a>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+          <div className="four-rectangle" />
+        </div>
+        <footer>
+          <div className="container footer-bg ">
+            <div className="container-1">
+              <div className="footer-group">
+                <div className="footer-left">
+                  <div>
+                    <div className="logo2">
+                      <img src="./svg/LOGO.svg" alt="no-pic" />
+                    </div>
+                  </div>
+                  <ul>
+                    <li>MERCURIUS COFFEE</li>
+                    <li>墨丘利咖啡</li>
+                    <li>TEL:0267-59-5557(營業時間8:00-17:00)</li>
+                    <li>Follow US </li>
+                    <div className="footer-icon">
+                      <li className="nav-icon-2">
+                        <a href="#/">
+                          <svg
+                            viewBox="-110 1 511 511.99996"
+                            xmlns="http://www.w3.org/2000/svg"
+                          >
+                            <path d="m180 512h-81.992188c-13.695312 0-24.835937-11.140625-24.835937-24.835938v-184.9375h-47.835937c-13.695313 0-24.835938-11.144531-24.835938-24.835937v-79.246094c0-13.695312 11.140625-24.835937 24.835938-24.835937h47.835937v-39.683594c0-39.347656 12.355469-72.824219 35.726563-96.804688 23.476562-24.089843 56.285156-36.820312 94.878906-36.820312l62.53125.101562c13.671875.023438 24.792968 11.164063 24.792968 24.835938v73.578125c0 13.695313-11.136718 24.835937-24.828124 24.835937l-42.101563.015626c-12.839844 0-16.109375 2.574218-16.808594 3.363281-1.152343 1.308593-2.523437 5.007812-2.523437 15.222656v31.351563h58.269531c4.386719 0 8.636719 1.082031 12.289063 3.121093 7.878906 4.402344 12.777343 12.726563 12.777343 21.722657l-.03125 79.246093c0 13.6875-11.140625 24.828125-24.835937 24.828125h-58.46875v184.941406c0 13.695313-11.144532 24.835938-24.839844 24.835938zm-76.8125-30.015625h71.632812v-193.195313c0-9.144531 7.441407-16.582031 16.582032-16.582031h66.726562l.027344-68.882812h-66.757812c-9.140626 0-16.578126-7.4375-16.578126-16.582031v-44.789063c0-11.726563 1.191407-25.0625 10.042969-35.085937 10.695313-12.117188 27.550781-13.515626 39.300781-13.515626l36.921876-.015624v-63.226563l-57.332032-.09375c-62.023437 0-100.566406 39.703125-100.566406 103.609375v53.117188c0 9.140624-7.4375 16.582031-16.578125 16.582031h-56.09375v68.882812h56.09375c9.140625 0 16.578125 7.4375 16.578125 16.582031zm163.0625-451.867187h.003906zm0 0" />
+                          </svg>
+                        </a>
+                      </li>
+                      <li className="nav-icon-2">
+                        <a href="#/">
+                          <svg
+                            viewBox="0 0 512.00096 512.00096"
+                            xmlns="http://www.w3.org/2000/svg"
+                          >
+                            <path d="m373.40625 0h-234.8125c-76.421875 0-138.59375 62.171875-138.59375 138.59375v234.816406c0 76.417969 62.171875 138.589844 138.59375 138.589844h234.816406c76.417969 0 138.589844-62.171875 138.589844-138.589844v-234.816406c0-76.421875-62.171875-138.59375-138.59375-138.59375zm108.578125 373.410156c0 59.867188-48.707031 108.574219-108.578125 108.574219h-234.8125c-59.871094 0-108.578125-48.707031-108.578125-108.574219v-234.816406c0-59.871094 48.707031-108.578125 108.578125-108.578125h234.816406c59.867188 0 108.574219 48.707031 108.574219 108.578125zm0 0" />
+                            <path d="m256 116.003906c-77.195312 0-139.996094 62.800782-139.996094 139.996094s62.800782 139.996094 139.996094 139.996094 139.996094-62.800782 139.996094-139.996094-62.800782-139.996094-139.996094-139.996094zm0 249.976563c-60.640625 0-109.980469-49.335938-109.980469-109.980469 0-60.640625 49.339844-109.980469 109.980469-109.980469 60.644531 0 109.980469 49.339844 109.980469 109.980469 0 60.644531-49.335938 109.980469-109.980469 109.980469zm0 0" />
+                            <path d="m399.34375 66.285156c-22.8125 0-41.367188 18.558594-41.367188 41.367188 0 22.8125 18.554688 41.371094 41.367188 41.371094s41.371094-18.558594 41.371094-41.371094-18.558594-41.367188-41.371094-41.367188zm0 52.71875c-6.257812 0-11.351562-5.09375-11.351562-11.351562 0-6.261719 5.09375-11.351563 11.351562-11.351563 6.261719 0 11.355469 5.089844 11.355469 11.351563 0 6.257812-5.09375 11.351562-11.355469 11.351562zm0 0" />
+                          </svg>
+                        </a>
+                      </li>
+                      <li className="nav-icon-2">
+                        <a href="#/">
+                          <svg
+                            viewBox="0 -45 512.00013 512"
+                            xmlns="http://www.w3.org/2000/svg"
+                          >
+                            <path d="m194.816406 422.710938c-85.453125 0-152.992187-14.929688-185.304687-40.953126l-.628907-.507812-.570312-.570312c-7.675781-7.679688-10.191406-17.753907-6.894531-27.636719l.300781-.820313c4-10.003906 13.800781-16.738281 24.421875-16.800781 21.859375-.378906 40.984375-2.984375 58.339844-8.042969-27.683594-12.875-46.914063-35.167968-58.355469-67.433594-3.847656-10.0625-.527344-21.21875 8.339844-27.871093 2.1875-1.644531 4.660156-2.886719 7.277344-3.71875-15.382813-17.757813-26.746094-37.964844-33.109376-59.335938l-.199218-.664062-.136719-.679688c-2.160156-10.808593 2.671875-21.921875 11.527344-26.707031 3.714843-2.132812 7.75-3.238281 11.800781-3.332031-4.367188-9.40625-7.542969-19.0625-9.425781-28.777344-5.226563-26.921875-.914063-53.910156 12.8125-80.214844l3.175781-6.351562c2.542969-5.082031 7.402344-8.652344 13.003906-9.5625 5.601563-.90625 11.34375.945312 15.355469 4.957031l5.785156 5.792969c45.703125 47.914062 86.640625 70.648437 157.417969 86.203125 3.160156-27.167969 14.90625-52.421875 33.855469-72.296875 22.550781-23.648438 52.664062-36.917969 84.792969-37.371094h.210937c23.441406 0 52.519531 13.382813 70.105469 22.820313 15.085937-4.9375 33.261718-12.582032 52.121094-20.664063 8.824218-4.140625 19.703124-2.2460938 26.640624 4.691406 6.800782 6.800781 8.6875 16.390625 5.078126 25.710938-1.371094 3.816406-2.925782 7.5625-4.65625 11.226562 2.582031 1.183594 4.945312 2.789063 6.941406 4.785157 6.035156 6.035156 8.550781 15.480468 6.40625 24.066406l-.230469.816406c-7.226563 23.289062-21.109375 42.257812-39.46875 54.164062-3.066406 163.285157-126.027344 295.078126-276.730469 295.078126zm-156.511718-57.675782c30.449218 17.226563 88.476562 27.648438 156.511718 27.648438 65.410156 0 127.136719-28.082032 173.804688-79.074219 47.050781-51.410156 72.960937-119.679687 72.960937-192.234375v-.816406c0-6.570313 3.617188-12.566406 9.4375-15.652344 11.808594-6.253906 21.371094-16.90625 27.589844-30.527344-6.414063 1.011719-12.933594-1.5625-16.929687-6.929687-4.644532-6.238281-4.695313-14.664063-.128907-20.957031 2.464844-3.398438 4.699219-6.933594 6.691407-10.589844-16.285157 6.839844-31.75 12.972656-45.175782 17.046875-4.878906 1.476562-10.316406.898437-14.773437-1.589844-23.902344-13.316406-46.164063-21.277344-59.585938-21.316406-49.527343.757812-89.796875 43.175781-89.796875 94.605469 0 5.316406-2.359375 10.300781-6.464844 13.679687-4.109374 3.375-9.453124 4.726563-14.671874 3.695313-81.609376-16.078126-129.96875-40.1875-180.257813-90.722657-7.207031 17.269531-9.175781 34.664063-5.84375 51.839844 3.378906 17.398437 12.367187 34.832031 25.996094 50.414063 5.179687 5.914062 5.867187 14.375 1.710937 21.050781-4.140625 6.652343-12.011718 9.761719-19.578125 7.734375-5.914062-1.585938-11.351562-3.667969-16.507812-6.34375 10.503906 22.816406 28.570312 43.917968 51.28125 59.480468 6.582031 4.511719 9.332031 12.921876 6.691406 20.453126-2.644531 7.542968-10.03125 12.398437-18.015625 11.804687-8.699219-.644531-16.40625-2.296875-23.5-5.082031 12.734375 25.933594 33.082031 40.203125 64.429688 45.65625 7.464843 1.300781 13.277343 7.195312 14.464843 14.667968 1.191407 7.472657-2.507812 14.878907-9.199219 18.429688-26.71875 14.164062-55.921874 21.765625-91.140624 23.628906zm0 0" />
+                          </svg>
+                        </a>
+                      </li>
+                      <li className="nav-icon-2">
+                        <a href="#/">
+                          <svg
+                            viewBox="0 -62 512.00199 512"
+                            xmlns="http://www.w3.org/2000/svg"
+                          >
+                            <path d="m334.808594 170.992188-113.113282-61.890626c-6.503906-3.558593-14.191406-3.425781-20.566406.351563-6.378906 3.78125-10.183594 10.460937-10.183594 17.875v122.71875c0 7.378906 3.78125 14.046875 10.117188 17.832031 3.308594 1.976563 6.976562 2.96875 10.652344 2.96875 3.367187 0 6.742187-.832031 9.847656-2.503906l113.117188-60.824219c6.714843-3.613281 10.90625-10.59375 10.9375-18.222656.027343-7.628906-4.113282-14.640625-10.808594-18.304687zm-113.859375 63.617187v-91.71875l84.539062 46.257813zm0 0" />
+                            <path d="m508.234375 91.527344-.023437-.234375c-.433594-4.121094-4.75-40.777344-22.570313-59.421875-20.597656-21.929688-43.949219-24.59375-55.179687-25.871094-.929688-.105469-1.78125-.203125-2.542969-.304688l-.894531-.09375c-67.6875-4.921874-169.910157-5.5937495-170.933594-5.59765575l-.089844-.00390625-.089844.00390625c-1.023437.00390625-103.246094.67578175-171.542968 5.59765575l-.902344.09375c-.726563.097657-1.527344.1875-2.398438.289063-11.101562 1.28125-34.203125 3.949219-54.859375 26.671875-16.972656 18.445312-21.878906 54.316406-22.382812 58.347656l-.058594.523438c-.152344 1.714844-3.765625 42.539062-3.765625 83.523437v38.3125c0 40.984375 3.613281 81.808594 3.765625 83.527344l.027344.257813c.433593 4.054687 4.746093 40.039062 22.484375 58.691406 19.367187 21.195312 43.855468 24 57.027344 25.507812 2.082031.238282 3.875.441406 5.097656.65625l1.183594.164063c39.082031 3.71875 161.617187 5.550781 166.8125 5.625l.15625.003906.15625-.003906c1.023437-.003907 103.242187-.675781 170.929687-5.597657l.894531-.09375c.855469-.113281 1.816406-.214843 2.871094-.324218 11.039062-1.171875 34.015625-3.605469 54.386719-26.019532 16.972656-18.449218 21.882812-54.320312 22.382812-58.347656l.058594-.523437c.152344-1.71875 3.769531-42.539063 3.769531-83.523438v-38.3125c-.003906-40.984375-3.617187-81.804687-3.769531-83.523437zm-26.238281 121.835937c0 37.933594-3.3125 77-3.625 80.585938-1.273438 9.878906-6.449219 32.574219-14.71875 41.5625-12.75 14.027343-25.847656 15.417969-35.410156 16.429687-1.15625.121094-2.226563.238282-3.195313.359375-65.46875 4.734375-163.832031 5.460938-168.363281 5.488281-5.082032-.074218-125.824219-1.921874-163.714844-5.441406-1.941406-.316406-4.039062-.558594-6.25-.808594-11.214844-1.285156-26.566406-3.042968-38.371094-16.027343l-.277344-.296875c-8.125-8.464844-13.152343-29.6875-14.429687-41.148438-.238281-2.710937-3.636719-42.238281-3.636719-80.703125v-38.3125c0-37.890625 3.304688-76.914062 3.625-80.574219 1.519532-11.636718 6.792969-32.957031 14.71875-41.574218 13.140625-14.453125 26.996094-16.054688 36.160156-17.113282.875-.101562 1.691407-.195312 2.445313-.292968 66.421875-4.757813 165.492187-5.464844 169.046875-5.492188 3.554688.023438 102.589844.734375 168.421875 5.492188.808594.101562 1.691406.203125 2.640625.3125 9.425781 1.074218 23.671875 2.699218 36.746094 16.644531l.121094.128906c8.125 8.464844 13.152343 30.058594 14.429687 41.75.226563 2.558594 3.636719 42.171875 3.636719 80.71875zm0 0" />
+                          </svg>
+                        </a>
+                      </li>
+                    </div>
+                  </ul>
+                </div>
+                <div className="footer-center">
+                  <div className="four-information">
+                    <div className="map">
+                      <ul>
+                        <li>
+                          {' '}
+                          <a href="#/">店家地圖</a>
+                        </li>
+                        <li>
+                          <a href="#/">店家畫面</a>
+                        </li>
+                      </ul>
+                    </div>
+                    <div className="coffee">
+                      <ul>
+                        <li>
+                          {' '}
+                          <a href="#/">地球儀展示</a>
+                        </li>
+                        <li>
+                          {' '}
+                          <a href="#/">台灣地圖展示</a>
+                        </li>
+                        <li>
+                          <a href="#/">屬性標籤分類</a>
+                        </li>
+                        <li>
+                          <a href="#/">產地/產品介紹</a>
+                        </li>
+                      </ul>
+                    </div>
+                    <div className="product">
+                      <ul>
+                        <li>
+                          <a href="#/"> 品牌周邊</a>
+                        </li>
+                        <li>
+                          <a href="#/" />
+                          咖啡
+                        </li>
+                        <li>
+                          <a href="#/">器材</a>
+                        </li>
+                        <li>
+                          <a href="#/">甜點</a>
+                        </li>
+                        <li>
+                          <a href="#/">購物車</a>
+                        </li>
+                      </ul>
+                    </div>
+                    <div className="partner">
+                      <ul>
+                        <li>
+                          <a href="#/"> 編輯店家資訊</a>
+                        </li>
+                        <li>
+                          <a href="#/">店家商品畫面</a>
+                        </li>
+                        <li>
+                          <a href="#/">放入購物車</a>
+                        </li>
+                        <li>
+                          <a href="#/">結帳</a>
+                        </li>
+                      </ul>
+                    </div>
+                  </div>
+                </div>
+                <div className="phonelogo2">
+                  <img src="./svg/phoneLOGO - .svg" alt="" />
+                </div>
+                <div className="phone-item">
+                  <div>墨丘利咖啡</div>
+                  <div>TEL:0267-59-5557(營業時間8:00-17:00)</div>
+                </div>
+                <div calss="footer-right">
+                  <div className="button-w mb-30">
+                    {' '}
+                    <a href="#/">聯絡我們</a>{' '}
+                  </div>
+                  <div className="button-w mb-30">
+                    <a href="#/">加入我們</a>
+                  </div>
+                  <div className="button-w mb-30">
+                    <a href="#/">關於我們</a>
+                  </div>
+                </div>
+                <div className="phone-footer-icon">
+                  <li className="nav-icon-2">
+                    <a href="#/">
+                      <svg
+                        viewBox="-110 1 511 511.99996"
+                        xmlns="http://www.w3.org/2000/svg"
+                      >
+                        <path d="m180 512h-81.992188c-13.695312 0-24.835937-11.140625-24.835937-24.835938v-184.9375h-47.835937c-13.695313 0-24.835938-11.144531-24.835938-24.835937v-79.246094c0-13.695312 11.140625-24.835937 24.835938-24.835937h47.835937v-39.683594c0-39.347656 12.355469-72.824219 35.726563-96.804688 23.476562-24.089843 56.285156-36.820312 94.878906-36.820312l62.53125.101562c13.671875.023438 24.792968 11.164063 24.792968 24.835938v73.578125c0 13.695313-11.136718 24.835937-24.828124 24.835937l-42.101563.015626c-12.839844 0-16.109375 2.574218-16.808594 3.363281-1.152343 1.308593-2.523437 5.007812-2.523437 15.222656v31.351563h58.269531c4.386719 0 8.636719 1.082031 12.289063 3.121093 7.878906 4.402344 12.777343 12.726563 12.777343 21.722657l-.03125 79.246093c0 13.6875-11.140625 24.828125-24.835937 24.828125h-58.46875v184.941406c0 13.695313-11.144532 24.835938-24.839844 24.835938zm-76.8125-30.015625h71.632812v-193.195313c0-9.144531 7.441407-16.582031 16.582032-16.582031h66.726562l.027344-68.882812h-66.757812c-9.140626 0-16.578126-7.4375-16.578126-16.582031v-44.789063c0-11.726563 1.191407-25.0625 10.042969-35.085937 10.695313-12.117188 27.550781-13.515626 39.300781-13.515626l36.921876-.015624v-63.226563l-57.332032-.09375c-62.023437 0-100.566406 39.703125-100.566406 103.609375v53.117188c0 9.140624-7.4375 16.582031-16.578125 16.582031h-56.09375v68.882812h56.09375c9.140625 0 16.578125 7.4375 16.578125 16.582031zm163.0625-451.867187h.003906zm0 0" />
+                      </svg>
+                    </a>
+                  </li>
+                  <li className="nav-icon-2">
+                    <a href="#/">
+                      <svg
+                        viewBox="0 0 512.00096 512.00096"
+                        xmlns="http://www.w3.org/2000/svg"
+                      >
+                        <path d="m373.40625 0h-234.8125c-76.421875 0-138.59375 62.171875-138.59375 138.59375v234.816406c0 76.417969 62.171875 138.589844 138.59375 138.589844h234.816406c76.417969 0 138.589844-62.171875 138.589844-138.589844v-234.816406c0-76.421875-62.171875-138.59375-138.59375-138.59375zm108.578125 373.410156c0 59.867188-48.707031 108.574219-108.578125 108.574219h-234.8125c-59.871094 0-108.578125-48.707031-108.578125-108.574219v-234.816406c0-59.871094 48.707031-108.578125 108.578125-108.578125h234.816406c59.867188 0 108.574219 48.707031 108.574219 108.578125zm0 0" />
+                        <path d="m256 116.003906c-77.195312 0-139.996094 62.800782-139.996094 139.996094s62.800782 139.996094 139.996094 139.996094 139.996094-62.800782 139.996094-139.996094-62.800782-139.996094-139.996094-139.996094zm0 249.976563c-60.640625 0-109.980469-49.335938-109.980469-109.980469 0-60.640625 49.339844-109.980469 109.980469-109.980469 60.644531 0 109.980469 49.339844 109.980469 109.980469 0 60.644531-49.335938 109.980469-109.980469 109.980469zm0 0" />
+                        <path d="m399.34375 66.285156c-22.8125 0-41.367188 18.558594-41.367188 41.367188 0 22.8125 18.554688 41.371094 41.367188 41.371094s41.371094-18.558594 41.371094-41.371094-18.558594-41.367188-41.371094-41.367188zm0 52.71875c-6.257812 0-11.351562-5.09375-11.351562-11.351562 0-6.261719 5.09375-11.351563 11.351562-11.351563 6.261719 0 11.355469 5.089844 11.355469 11.351563 0 6.257812-5.09375 11.351562-11.355469 11.351562zm0 0" />
+                      </svg>
+                    </a>
+                  </li>
+                  <li className="nav-icon-2">
+                    <a href="#/">
+                      <svg
+                        viewBox="0 -45 512.00013 512"
+                        xmlns="http://www.w3.org/2000/svg"
+                      >
+                        <path d="m194.816406 422.710938c-85.453125 0-152.992187-14.929688-185.304687-40.953126l-.628907-.507812-.570312-.570312c-7.675781-7.679688-10.191406-17.753907-6.894531-27.636719l.300781-.820313c4-10.003906 13.800781-16.738281 24.421875-16.800781 21.859375-.378906 40.984375-2.984375 58.339844-8.042969-27.683594-12.875-46.914063-35.167968-58.355469-67.433594-3.847656-10.0625-.527344-21.21875 8.339844-27.871093 2.1875-1.644531 4.660156-2.886719 7.277344-3.71875-15.382813-17.757813-26.746094-37.964844-33.109376-59.335938l-.199218-.664062-.136719-.679688c-2.160156-10.808593 2.671875-21.921875 11.527344-26.707031 3.714843-2.132812 7.75-3.238281 11.800781-3.332031-4.367188-9.40625-7.542969-19.0625-9.425781-28.777344-5.226563-26.921875-.914063-53.910156 12.8125-80.214844l3.175781-6.351562c2.542969-5.082031 7.402344-8.652344 13.003906-9.5625 5.601563-.90625 11.34375.945312 15.355469 4.957031l5.785156 5.792969c45.703125 47.914062 86.640625 70.648437 157.417969 86.203125 3.160156-27.167969 14.90625-52.421875 33.855469-72.296875 22.550781-23.648438 52.664062-36.917969 84.792969-37.371094h.210937c23.441406 0 52.519531 13.382813 70.105469 22.820313 15.085937-4.9375 33.261718-12.582032 52.121094-20.664063 8.824218-4.140625 19.703124-2.2460938 26.640624 4.691406 6.800782 6.800781 8.6875 16.390625 5.078126 25.710938-1.371094 3.816406-2.925782 7.5625-4.65625 11.226562 2.582031 1.183594 4.945312 2.789063 6.941406 4.785157 6.035156 6.035156 8.550781 15.480468 6.40625 24.066406l-.230469.816406c-7.226563 23.289062-21.109375 42.257812-39.46875 54.164062-3.066406 163.285157-126.027344 295.078126-276.730469 295.078126zm-156.511718-57.675782c30.449218 17.226563 88.476562 27.648438 156.511718 27.648438 65.410156 0 127.136719-28.082032 173.804688-79.074219 47.050781-51.410156 72.960937-119.679687 72.960937-192.234375v-.816406c0-6.570313 3.617188-12.566406 9.4375-15.652344 11.808594-6.253906 21.371094-16.90625 27.589844-30.527344-6.414063 1.011719-12.933594-1.5625-16.929687-6.929687-4.644532-6.238281-4.695313-14.664063-.128907-20.957031 2.464844-3.398438 4.699219-6.933594 6.691407-10.589844-16.285157 6.839844-31.75 12.972656-45.175782 17.046875-4.878906 1.476562-10.316406.898437-14.773437-1.589844-23.902344-13.316406-46.164063-21.277344-59.585938-21.316406-49.527343.757812-89.796875 43.175781-89.796875 94.605469 0 5.316406-2.359375 10.300781-6.464844 13.679687-4.109374 3.375-9.453124 4.726563-14.671874 3.695313-81.609376-16.078126-129.96875-40.1875-180.257813-90.722657-7.207031 17.269531-9.175781 34.664063-5.84375 51.839844 3.378906 17.398437 12.367187 34.832031 25.996094 50.414063 5.179687 5.914062 5.867187 14.375 1.710937 21.050781-4.140625 6.652343-12.011718 9.761719-19.578125 7.734375-5.914062-1.585938-11.351562-3.667969-16.507812-6.34375 10.503906 22.816406 28.570312 43.917968 51.28125 59.480468 6.582031 4.511719 9.332031 12.921876 6.691406 20.453126-2.644531 7.542968-10.03125 12.398437-18.015625 11.804687-8.699219-.644531-16.40625-2.296875-23.5-5.082031 12.734375 25.933594 33.082031 40.203125 64.429688 45.65625 7.464843 1.300781 13.277343 7.195312 14.464843 14.667968 1.191407 7.472657-2.507812 14.878907-9.199219 18.429688-26.71875 14.164062-55.921874 21.765625-91.140624 23.628906zm0 0" />
+                      </svg>
+                    </a>
+                  </li>
+                  <li className="nav-icon-2">
+                    <a href="#/">
+                      <svg
+                        viewBox="0 -62 512.00199 512"
+                        xmlns="http://www.w3.org/2000/svg"
+                      >
+                        <path d="m334.808594 170.992188-113.113282-61.890626c-6.503906-3.558593-14.191406-3.425781-20.566406.351563-6.378906 3.78125-10.183594 10.460937-10.183594 17.875v122.71875c0 7.378906 3.78125 14.046875 10.117188 17.832031 3.308594 1.976563 6.976562 2.96875 10.652344 2.96875 3.367187 0 6.742187-.832031 9.847656-2.503906l113.117188-60.824219c6.714843-3.613281 10.90625-10.59375 10.9375-18.222656.027343-7.628906-4.113282-14.640625-10.808594-18.304687zm-113.859375 63.617187v-91.71875l84.539062 46.257813zm0 0" />
+                        <path d="m508.234375 91.527344-.023437-.234375c-.433594-4.121094-4.75-40.777344-22.570313-59.421875-20.597656-21.929688-43.949219-24.59375-55.179687-25.871094-.929688-.105469-1.78125-.203125-2.542969-.304688l-.894531-.09375c-67.6875-4.921874-169.910157-5.5937495-170.933594-5.59765575l-.089844-.00390625-.089844.00390625c-1.023437.00390625-103.246094.67578175-171.542968 5.59765575l-.902344.09375c-.726563.097657-1.527344.1875-2.398438.289063-11.101562 1.28125-34.203125 3.949219-54.859375 26.671875-16.972656 18.445312-21.878906 54.316406-22.382812 58.347656l-.058594.523438c-.152344 1.714844-3.765625 42.539062-3.765625 83.523437v38.3125c0 40.984375 3.613281 81.808594 3.765625 83.527344l.027344.257813c.433593 4.054687 4.746093 40.039062 22.484375 58.691406 19.367187 21.195312 43.855468 24 57.027344 25.507812 2.082031.238282 3.875.441406 5.097656.65625l1.183594.164063c39.082031 3.71875 161.617187 5.550781 166.8125 5.625l.15625.003906.15625-.003906c1.023437-.003907 103.242187-.675781 170.929687-5.597657l.894531-.09375c.855469-.113281 1.816406-.214843 2.871094-.324218 11.039062-1.171875 34.015625-3.605469 54.386719-26.019532 16.972656-18.449218 21.882812-54.320312 22.382812-58.347656l.058594-.523437c.152344-1.71875 3.769531-42.539063 3.769531-83.523438v-38.3125c-.003906-40.984375-3.617187-81.804687-3.769531-83.523437zm-26.238281 121.835937c0 37.933594-3.3125 77-3.625 80.585938-1.273438 9.878906-6.449219 32.574219-14.71875 41.5625-12.75 14.027343-25.847656 15.417969-35.410156 16.429687-1.15625.121094-2.226563.238282-3.195313.359375-65.46875 4.734375-163.832031 5.460938-168.363281 5.488281-5.082032-.074218-125.824219-1.921874-163.714844-5.441406-1.941406-.316406-4.039062-.558594-6.25-.808594-11.214844-1.285156-26.566406-3.042968-38.371094-16.027343l-.277344-.296875c-8.125-8.464844-13.152343-29.6875-14.429687-41.148438-.238281-2.710937-3.636719-42.238281-3.636719-80.703125v-38.3125c0-37.890625 3.304688-76.914062 3.625-80.574219 1.519532-11.636718 6.792969-32.957031 14.71875-41.574218 13.140625-14.453125 26.996094-16.054688 36.160156-17.113282.875-.101562 1.691407-.195312 2.445313-.292968 66.421875-4.757813 165.492187-5.464844 169.046875-5.492188 3.554688.023438 102.589844.734375 168.421875 5.492188.808594.101562 1.691406.203125 2.640625.3125 9.425781 1.074218 23.671875 2.699218 36.746094 16.644531l.121094.128906c8.125 8.464844 13.152343 30.058594 14.429687 41.75.226563 2.558594 3.636719 42.171875 3.636719 80.71875zm0 0" />
+                      </svg>
+                    </a>
+                  </li>
+                </div>
+              </div>
+              <div className="copyright">
+                Copyright © 2021 MERCURIUS - Coffee Server &amp; Design.
+              </div>
+            </div>
+          </div>
+        </footer>
+      </div>
+      );
+    </>
+  )
+}
+
+export default App
