+++ src/pages/index/index.html
@@ -14,6 +14,7 @@
 
 <body>
     <div class="container">
+
         <div class="nav ">
 
             <div class="icon-s-none ">
@@ -348,7 +349,11 @@
             </div>
 
         </div>
+    </div>
 
+
+
+    <div class="container">
         <div class="herro-section">
             <div class="onepage-pic">
                 <img src="./img/herrosection.jpg" alt="no-pic">
@@ -734,12 +739,32 @@
             </div>
 
         </div>
-
-
         <div class='four-rectangle'></div>
 
-
     </div>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
     <footer>
         <div class="container footer-bg ">
